backend/requirements.txt
langchain-community
langchain-text-splitters
langchain-huggingface
sentence-transformers
langchain-google-genai
python-multipart
langchain-chroma
sqlalchemy
PyJWT
fastapi
uvicorn
pymysql